no message
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@77 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/include/MIME_manager.h
@@ -47,13 +47,13 @@ class MIME_Manager
 		mime_record* next;
 	};
 	mime_record *data;
-	DWORD numMimeTypesLoaded;
+	u_long numMimeTypesLoaded;
 public:
 	MIME_Manager();
 	VOID addRecord(MIME_Manager::mime_record);
 	VOID removeAllRecords();
 	VOID removeRecord(char*);
-	DWORD getNumMIMELoaded();
+	u_long getNumMIMELoaded();
 	int load(char *filename);
 	int getMIME(char* ext,char *dest,char *dest2);
 	VOID clean();
