+++ myserverweb/source/sockets.cpp
@@ -119,10 +119,10 @@ int ms_recv(MYSERVER_SOCKET s,char* buffer,int len,int flags)
 #endif
 }
 
-DWORD bytesToRead(MYSERVER_SOCKET c)
+u_long bytesToRead(MYSERVER_SOCKET c)
 {
 #ifdef WIN32
-	DWORD nBytesToRead;
+	u_long nBytesToRead;
 	ms_ioctlsocket(c,FIONREAD,&nBytesToRead);
 	return nBytesToRead;
 #endif
